feat(ui): move Clear/New batch action beside Run
