+++ src/components/StatusBar.test.tsx
@@ -1,6 +1,5 @@
-import { render, screen, within } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
-import { beforeEach, describe, expect, it, vi } from "vitest";
+import { render, screen } from "@testing-library/react";
+import { beforeEach, describe, expect, it } from "vitest";
 
 import { resetJobStore, useJobStore } from "@/store/jobStore";
 
@@ -124,29 +123,15 @@ describe("StatusBar", () => {
 });
 
 // ---------------------------------------------------------------------------
-// Reset slot — visibility
+// Reset action — relocated to RunControls (regression guard)
 // ---------------------------------------------------------------------------
 
-describe("StatusBar Reset slot – visibility", () => {
+describe("StatusBar – reset action moved out", () => {
   beforeEach(() => resetJobStore());
 
-  it("does not show Reset when items is empty (not running)", () => {
-    // default state: no items, not running
-    render(<StatusBar />);
-    expect(
-      screen.queryByRole("button", { name: /clear|new batch/i }),
-    ).toBeNull();
-  });
-
-  it("does not show Reset when items is empty and running", () => {
-    useJobStore.setState({ running: true });
-    render(<StatusBar />);
-    expect(
-      screen.queryByRole("button", { name: /clear|new batch/i }),
-    ).toBeNull();
-  });
-
-  it("shows Reset when hasItems and not running (no summary)", () => {
+  // The Clear / New batch action moved to RunControls (beside Run). The slim
+  // footer must no longer render it, in any queue/summary state.
+  it("does not render a Clear/New batch button when items are queued", () => {
     useJobStore.setState({
       draft: {
         items: [ITEM],
@@ -160,29 +145,12 @@ describe("StatusBar Reset slot – visibility", () => {
       summary: null,
     });
     render(<StatusBar />);
-    expect(screen.getByRole("button", { name: /clear/i })).toBeTruthy();
-  });
-
-  it("does not show Reset when hasItems but running", () => {
-    useJobStore.setState({
-      draft: {
-        items: [ITEM],
-        namingTemplate: null,
-        startNumber: 1,
-        outputDir: null,
-        outputMode: "zip",
-        conflictPolicy: "autoRename",
-      },
-      running: true,
-      summary: null,
-    });
-    render(<StatusBar />);
     expect(
       screen.queryByRole("button", { name: /clear|new batch/i }),
     ).toBeNull();
   });
 
-  it("shows Reset after job finishes (summary set, hasItems, not running)", () => {
+  it("does not render a Clear/New batch button after a job finishes", () => {
     useJobStore.setState({
       draft: {
         items: [ITEM],
@@ -196,237 +164,8 @@ describe("StatusBar Reset slot – visibility", () => {
       summary: { succeeded: [1], cancelled: [], failed: [], results: [] },
     });
     render(<StatusBar />);
-    expect(screen.getByRole("button", { name: /new batch/i })).toBeTruthy();
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Reset slot — label
-// ---------------------------------------------------------------------------
-
-describe("StatusBar Reset slot – label", () => {
-  beforeEach(() => resetJobStore());
-
-  it("labels the button 'Clear' when summary is null", () => {
-    useJobStore.setState({
-      draft: {
-        items: [ITEM],
-        namingTemplate: null,
-        startNumber: 1,
-        outputDir: null,
-        outputMode: "zip",
-        conflictPolicy: "autoRename",
-      },
-      running: false,
-      summary: null,
-    });
-    render(<StatusBar />);
-    expect(screen.getByRole("button", { name: /^clear$/i })).toBeTruthy();
-  });
-
-  it("labels the button 'New batch' when summary is not null", () => {
-    useJobStore.setState({
-      draft: {
-        items: [ITEM],
-        namingTemplate: null,
-        startNumber: 1,
-        outputDir: null,
-        outputMode: "zip",
-        conflictPolicy: "autoRename",
-      },
-      running: false,
-      summary: {
-        succeeded: [],
-        cancelled: [],
-        failed: [{ taskId: 1, reason: "error" }],
-        results: [],
-      },
-    });
-    render(<StatusBar />);
-    expect(screen.getByRole("button", { name: /^new batch$/i })).toBeTruthy();
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Reset slot — ConfirmDialog flow
-// ---------------------------------------------------------------------------
-
-describe("StatusBar Reset slot – confirm dialog", () => {
-  beforeEach(() => resetJobStore());
-
-  it("opens the dialog when the Reset button is clicked", async () => {
-    useJobStore.setState({
-      draft: {
-        items: [ITEM],
-        namingTemplate: null,
-        startNumber: 1,
-        outputDir: null,
-        outputMode: "zip",
-        conflictPolicy: "autoRename",
-      },
-      running: false,
-      summary: null,
-    });
-    const user = userEvent.setup();
-    render(<StatusBar />);
-    expect(screen.queryByRole("dialog")).toBeNull();
-    await user.click(screen.getByRole("button", { name: /^clear$/i }));
-    expect(screen.getByRole("dialog")).toBeTruthy();
-  });
-
-  it("calls reset() once when the user confirms the dialog (Clear flow)", async () => {
-    const reset = vi.fn().mockResolvedValue(undefined);
-    useJobStore.setState({
-      draft: {
-        items: [ITEM],
-        namingTemplate: null,
-        startNumber: 1,
-        outputDir: null,
-        outputMode: "zip",
-        conflictPolicy: "autoRename",
-      },
-      running: false,
-      summary: null,
-      reset,
-    });
-    const user = userEvent.setup();
-    render(<StatusBar />);
-    await user.click(screen.getByRole("button", { name: /^clear$/i }));
-    // The confirm button inside the dialog: scope to dialog to avoid ambiguity.
-    const dialog = screen.getByRole("dialog");
-    await user.click(within(dialog).getByRole("button", { name: /^clear$/i }));
-    expect(reset).toHaveBeenCalledTimes(1);
-  });
-
-  it("does not call reset() when the user cancels the dialog", async () => {
-    const reset = vi.fn().mockResolvedValue(undefined);
-    useJobStore.setState({
-      draft: {
-        items: [ITEM],
-        namingTemplate: null,
-        startNumber: 1,
-        outputDir: null,
-        outputMode: "zip",
-        conflictPolicy: "autoRename",
-      },
-      running: false,
-      summary: null,
-      reset,
-    });
-    const user = userEvent.setup();
-    render(<StatusBar />);
-    await user.click(screen.getByRole("button", { name: /^clear$/i }));
-    await user.click(
-      within(screen.getByRole("dialog")).getByRole("button", {
-        name: /^cancel$/i,
-      }),
-    );
-    expect(reset).not.toHaveBeenCalled();
-  });
-
-  it("calls reset() once when confirming the 'New batch' dialog", async () => {
-    const reset = vi.fn().mockResolvedValue(undefined);
-    useJobStore.setState({
-      draft: {
-        items: [ITEM],
-        namingTemplate: null,
-        startNumber: 1,
-        outputDir: null,
-        outputMode: "zip",
-        conflictPolicy: "autoRename",
-      },
-      running: false,
-      summary: { succeeded: [1], cancelled: [], failed: [], results: [] },
-      reset,
-    });
-    const user = userEvent.setup();
-    render(<StatusBar />);
-    await user.click(screen.getByRole("button", { name: /^new batch$/i }));
-    // Scope to the dialog to avoid ambiguity with the footer Reset button.
-    const dialog = screen.getByRole("dialog");
-    await user.click(
-      within(dialog).getByRole("button", { name: /^new batch$/i }),
-    );
-    expect(reset).toHaveBeenCalledTimes(1);
-  });
-
-  it("does not call reset() when cancelling the 'New batch' dialog", async () => {
-    const reset = vi.fn().mockResolvedValue(undefined);
-    useJobStore.setState({
-      draft: {
-        items: [ITEM],
-        namingTemplate: null,
-        startNumber: 1,
-        outputDir: null,
-        outputMode: "zip",
-        conflictPolicy: "autoRename",
-      },
-      running: false,
-      summary: { succeeded: [1], cancelled: [], failed: [], results: [] },
-      reset,
-    });
-    const user = userEvent.setup();
-    render(<StatusBar />);
-    await user.click(screen.getByRole("button", { name: /^new batch$/i }));
-    await user.click(
-      within(screen.getByRole("dialog")).getByRole("button", {
-        name: /^cancel$/i,
-      }),
-    );
-    expect(reset).not.toHaveBeenCalled();
-  });
-
-  it("closes the dialog after confirming", async () => {
-    const reset = vi.fn().mockResolvedValue(undefined);
-    useJobStore.setState({
-      draft: {
-        items: [ITEM],
-        namingTemplate: null,
-        startNumber: 1,
-        outputDir: null,
-        outputMode: "zip",
-        conflictPolicy: "autoRename",
-      },
-      running: false,
-      summary: null,
-      reset,
-    });
-    const user = userEvent.setup();
-    render(<StatusBar />);
-    await user.click(screen.getByRole("button", { name: /^clear$/i }));
-    expect(screen.getByRole("dialog")).toBeTruthy();
-    await user.click(
-      within(screen.getByRole("dialog")).getByRole("button", {
-        name: /^clear$/i,
-      }),
-    );
-    expect(screen.queryByRole("dialog")).toBeNull();
-  });
-
-  it("closes the dialog after cancelling", async () => {
-    const reset = vi.fn().mockResolvedValue(undefined);
-    useJobStore.setState({
-      draft: {
-        items: [ITEM],
-        namingTemplate: null,
-        startNumber: 1,
-        outputDir: null,
-        outputMode: "zip",
-        conflictPolicy: "autoRename",
-      },
-      running: false,
-      summary: null,
-      reset,
-    });
-    const user = userEvent.setup();
-    render(<StatusBar />);
-    await user.click(screen.getByRole("button", { name: /^clear$/i }));
-    expect(screen.getByRole("dialog")).toBeTruthy();
-    await user.click(
-      within(screen.getByRole("dialog")).getByRole("button", {
-        name: /^cancel$/i,
-      }),
-    );
-    expect(screen.queryByRole("dialog")).toBeNull();
+    expect(
+      screen.queryByRole("button", { name: /clear|new batch/i }),
+    ).toBeNull();
   });
 });
